requirements.txt
PythonTurtle
requests
pandas
pyarrow
python-dotenv
pathlib
pyarrow == 12.0.1